rm history plugin
part of #126
